+++ docs/news.rst
@@ -7,6 +7,7 @@ next (unreleased)
 Downward Lab
 ^^^
 * Add buttons for revealing and hiding all tables in an AbsoluteReport section (Jendrik Seipp).
+* Label more ticks in relative scatter plots (Jendrik Seipp).
 
 
 v8.8 (2026-01-17)